pyproject.toml
[project]
name = "ask-samuel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.7",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "mistralai>=1.9.2",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.3",
]


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "C4",     # flake8-comprehensions
    "B",      # flake8-bugbear
    "RUF",    # Ruff-specific rules
    "N",      # pep8-naming
    "SIM",    # flake8-simplify
    "PTH",    # flake8-use-pathlib
    "ANN001", # Checks that function arguments have type annotations.
    "ANN201", # Missing return type annotation for public function {name}
    "DTZ",    # flake8-datetimez: checks for naive datetime usage
    "T201",   # Report print statements
]

ignore = ["E501"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"


# --- Per-File Ignores ---
[tool.ruff.lint.per-file-ignores]
# Allow unused imports in __init__.py files for exposing package APIs.
"**/__init__.py" = ["F401"]


[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
